fix a bug with loading dataset.appConfig setting
